+++ src/scenes/end.ts
@@ -1,0 +1,15 @@
+import * as Phaser from "./../../node_modules/phaser/dist/phaser.js";
+
+export class EndScene extends Phaser.Scene {
+    constructor() {
+        super();
+    }
+
+    preload() {
+
+    }
+
+    create() {
+
+    }
+}