+++ src/scenes/load.ts
@@ -1,0 +1,15 @@
+import * as Phaser from "./../../node_modules/phaser/dist/phaser.js";
+
+export class LoadScene extends Phaser.Scene {
+    constructor() {
+        super();
+    }
+
+    preload() {
+        console.log("in load preload!")
+    }
+
+    create() {
+
+    }
+}